refactor(KonnectorAction): Extract helpers
findMatchingBrandWithoutTrigger and hasUrls are extracted to helpers.js.
The goal is to make match.js as little and as reusable as possible in
customizations
